* master: (320 commits)
  Orientation-aware key glyphs (#4757)
  Fix warning in geom_violin with draw_quantiles (#4654)
  Fix misalignment in geom_dotplot when stackratio != 1 and stackdir != "up" (#4734)
  Replace coord_equal in scale-identity.r (#4759)
  Unified message format in `geom_smooth()` (#4634)
  Add parentheses to mentions of binned_scale in scale-alpha and scale-viridis documentation (#4735)
  Update geom-boxplot.r (#4744)
  Remove unneeded backslash from diamonds docs (#4711)
  Re-document
  Warn on unsupported geoms in annotate() (#4721)
  Re-document
  Add Trump to presidential terms dataset (#4702)
  Corrects bins / binwidth override documentation (#4720)
  Update infrastructure to new best practices (#4748)
  remove stringsAsFactor for the mapped_discrete as.data.frame method (#4750)
  Workflow updates (#4747)
  Ensure output is numeric even if ifelse clause is NA (#4692)
  Add non_missing_aes to geom_tile (#4683)
  Pass on binwidth and height to geom (#4671)
  Check for range == NULL - happens if no data has been added to plot (#4682)
  ...

+++ R/geom-.r
@@ -3,7 +3,7 @@ NULL
 
 #' @section Geoms:
 #'
-#' All `geom_*` functions (like `geom_point`) return a layer that
+#' All `geom_*()` functions (like `geom_point()`) return a layer that
 #' contains a `Geom*` object (like `GeomPoint`). The `Geom*`
 #' object is responsible for rendering the data in the plot.
 #'
@@ -14,9 +14,7 @@ NULL
 #' Compared to `Stat` and `Position`, `Geom` is a little
 #' different because the execution of the setup and compute functions is
 #' split up. `setup_data` runs before position adjustments, and
-#' `draw_layer` is not run until render time, much later. This
-#' means there is no `setup_params` because it's hard to communicate
-#' the changes.
+#' `draw_layer()` is not run until render time, much later.
 #'
 #' To create a new type of Geom object, you typically will want to
 #' override one or more of the following:
@@ -210,7 +208,7 @@ NULL
 .stroke <- 96 / 25.4
 
 check_aesthetics <- function(x, n) {
-  ns <- vapply(x, length, numeric(1))
+  ns <- vapply(x, length, integer(1))
   good <- ns == 1L | ns == n
 
   if (all(good)) {
